Update run_sid.sh to avoid confusions on how to reproduce our reported results
As explicitly stated in the paper and illustrated in the figures, the default setting for reproducing the reported results was a duration of 500 million (--duration 500), except for CIFAR-10 conditional under alpha = 1.2, where it was increased to 800, and ImageNet 64×64 under alpha = 1.2, where it was raised to 1000. Additionally, we have released all our checkpoints for further verification. Setting --duration 100 will produce very competitive results, as shown in Figures 2–6, but will not match the numbers reported in Tables 1–4.